Trust the emails forwarded by Helsinki profile
As with new information, Helsinki profile verifies the email
addresses themselves, so we can trust the emails that it forwards,
no matter what the underlying service is.

In the future this feature might come handy again, which is why
it is left in place. E.g. in case Helsinki profile implements
alternative authentication flows.