Prepare for Keycloak switch and implement the documented changes
From the documentation regarding Keycloak it seems there is only a
single authentication server which provides the authentication
tokens for both the user and the backend.

There are also some notes about the API scopes that with Keycloak
we do not need to provide the service URLs within the API scopes.