Rename GDPR API to Profile API for sanity
It appears we have originally incorrectly named the Profile
GraphQL API as "GDPR API". This is incorrect. The GraphQL API to
fetch profile details is named "Profile API".

To avoid confusion in the future, we rename it also in the code
because the GDPR API is different (managing the user's GDPR
requests). The end application does not need to call the GDPR API
itself, it is handled by the profile application itself. The end
application only needs to provide its own GDPR API endpoints for
the GDPR requests (get details, remove details, etc.).